refactor: split frontend lib.rs into focused modules (#202)
The crate root can't become a directory, so extract three sibling modules from
the 1,519-line frontend/src/lib.rs: styles.rs (the STYLES CSS const), routes.rs
(the Route enum + route-wrapper components), and contexts.rs (server-url /
search-query / current-user context helpers). lib.rs keeps App, both
ScreenLayout variants, and re-exports the extracted items so omnibus_frontend::
{Route, STYLES, use_current_user, ...} paths stay valid. lib.rs drops to 218
lines. No behavior change; feature-gating preserved verbatim.
